refactor vault processor
