+++ crates/indexer/src/models/vault_models/mod.rs
@@ -1,6 +1,8 @@
 // Copyright © Aptos Foundation
 
 pub mod rebase;
+pub mod user_infos;
 pub mod vault_activities;
 pub mod vault_events;
-pub mod vault_resources;
+pub mod vault_utils;
+pub mod vaults;
